add single post view

+++ src/Controllers/PostsController.php
@@ -1,9 +1,7 @@
 <?php
 
 namespace App\Controllers;
-
 use App\Models\Post;
-
 class PostsController
 {
     public function index(){
@@ -22,27 +20,58 @@ class PostsController
         $post->save();
         redirect('/admin/posts');
     }
-    
+
     public function edit(){
+        if (!isset($_GET['id']) || !is_numeric($_GET['id'])) {
+            redirect('/admin/posts');
+            return;
+        }
+
         $post = Post::find($_GET['id']);
+        if (!$post) {
+            redirect('/admin/posts');
+            return;
+        }
+
         view('posts/edit', compact('post'));
     }
 
     public function update(){
+
+        if (!isset($_GET['id']) || !is_numeric($_GET['id'])) {
+            redirect('/admin/posts');
+            return;
+        }
+
         $post = Post::find($_GET['id']);
+        if (!$post) {
+            redirect('/admin/posts');
+            return;
+        }
+
         $post->title = $_POST['title'];
         $post->body = $_POST['body'];
         $post->save();
         redirect('/admin/posts');
     }
+
     public function destroy(){
+        if (!isset($_GET['id']) || !is_numeric($_GET['id'])) {
+            redirect('/admin/posts');
+            return;
+        }
+
         $post = Post::find($_GET['id']);
-        if($post){
+        if ($post) {
             $post->delete();
         }
         redirect('/admin/posts');
     }
-    public function show(){ //tööööö, uus route, uus vaade, väike tabel andmetega, nagu edit, aga formi ja inputi asemel tabel
 
+    public function show(){ //SINCE WHEN DO WE HAVE HOMEWORK
+        $post = Post::find($_GET['id']);
+        view('posts/show', compact('post'));
+        // redirect('/admin/posts');
     }
+
 }
